Struct state_synchronizer::StateSynchronizer

pub struct StateSynchronizer { /* fields omitted */ }

Methods

impl StateSynchronizer

pub fn bootstrap(
network: Vec<(PeerId, StateSynchronizerSender, StateSynchronizerEvents)>,
state_sync_to_mempool_sender: Sender<CommitNotification>,
storage: Arc<dyn DbReader>,
executor: Box<dyn ChunkExecutor>,
config: &NodeConfig,
reconfig_event_subscriptions: Vec<ReconfigSubscription>
) -> Self

Setup state synchronizer. spawns coordinator and downloader routines on executor

pub fn bootstrap_with_executor_proxy<E: ExecutorProxyTrait + 'static>(
runtime: Runtime,
network: Vec<(PeerId, StateSynchronizerSender, StateSynchronizerEvents)>,
state_sync_to_mempool_sender: Sender<CommitNotification>,
role: RoleType,
waypoint: Option<Waypoint>,
state_sync_config: &StateSyncConfig,
upstream_config: UpstreamConfig,
executor_proxy: E
) -> Self

pub fn create_client(&self) -> Arc<StateSyncClient>

pub async fn wait_until_initialized<'_>(&'_ self) -> Result<()>

The function returns a future that is fulfilled when the state synchronizer is caught up with the waypoint specified in the local config.